==> 15_crqa_analysis.Rmd <==
---
title: "CRQA Final Results"
author: "Adam Stone, PhD"
date: '`r format(Sys.Date(), "%m-%d-%Y")`'
output:
  html_notebook:
    code_folding: hide
    highlight: tango
    theme: yeti
    toc: yes
    toc_depth: 2
    toc_float: true
---

```{r Load Libraries, message=FALSE, warning=FALSE}
# Load libraries
library(tidyverse)
library(janitor)
library(lme4)
library(broom)
library(broom.mixed)
library(sjPlot)
library(assertr)
library(skimr)
library(hrbrthemes)
library(crqa)
```


OK! After messy code in 12_, 13_, and 14_ (should clean it up, eek), we have CRQA results all over. Cool. Let's load 'em all and smoosh up all the input data and output variables in one neat data frame. 

```{r Loading & Tidying Data, message=FALSE, warning=FALSE}

# This are the data frames (with lists, etc.) used for finding the optimal
# radius for each person's eye gaze, against both the right hand and the left
# hand. First, we took all data associated with each story x direction x hand,
# and found the optimal embeddim and delay parameters for each via the
# optimizeParam() function. Then we ran those optimal parameters in CRQAs with a
# radius range of 5-60 ("test radius") to get associated rr_values. The target
# rr_value was 5, and the test radius that produced the target rr_value was then
# used for the final CRQA.
crqa_inputdata_r <- readRDS("crqa_rhand.RDS") %>%
  add_column(hand = "right") %>%
  rename(hand_y = rhand_y,
         hand_len = rhand_len,
         radius = test_radius)
  
crqa_inputdata_l <- readRDS("crqa_lhand.RDS") %>%
  add_column(hand = "left") %>%
  rename(hand_y = lhand_y,
         hand_len = lhand_len,
         radius = test_radius)

input_data <- bind_rows(crqa_inputdata_r,
                        crqa_inputdata_l)

rm('crqa_inputdata_r')
rm('crqa_inputdata_l')


# These are the CRQA results for each participant x story x direction x hand,
# using the test radius found above.
crqa_results_r <- read_csv("crqa_results_rhand.csv") %>%
  add_column(hand = "right") %>%
  rename_at(vars(rhand_rr:rhand_tt), ~ str_remove(.x, "rhand_")) %>%
  rename(radius = optimized_r)

crqa_results_l <- read_csv("crqa_results_lhand.csv") %>%
  add_column(hand = "left") %>%
  rename_at(vars(lhand_rr:lhand_tt), ~ str_remove(.x, "lhand_")) %>%
  rename(radius = optimized_r)

results <- bind_rows(crqa_results_r,
                     crqa_results_l)

rm('crqa_results_r')
rm('crqa_results_l')

# Now combine everything in one neat data frame
final <- results %>%
  left_join(input_data, by = c("name",
                               "maingroup",
                               "story",
                               "direction",
                               "embeddim",
                               "delay",
                               "radius",
                               "hand")) %>%
  verify(rr == rr_values) %>%
  mutate_at(vars(name, maingroup, story, direction, hand), as.factor) %>%
  select(-rr_values)

rm('results')
rm('input_data')

# Print summaries
skim_with(numeric = list(p25 = NULL,
                         p75 = NULL),
          integer = list(p25 = NULL,
                         p75 = NULL))
skim(final)
```

# Stats Analysis: Radius Values

Because we fixed the %REC (`RR`) to 5.0, that means the radius varies for each row. See below:

> If the data sets cannot be reasonably fitted with a single radius parameter...one can adjust the radius for each data set individually in order to keep the percentage of recurrence constant across all data sets (e.g., fixed percent recurrence of %REC = 5% for each data set). Of course, in this case, %REC needs to be omitted from the inferential statistical analysis, as it should be very similar (or the same) across pairings in all samples, but the other measures, such as %DET, ADL, and MDL can be still analyzed (as a stand-in for %REC, one can instead include analysis of the radius parameter r that is now different across pairs of groups of time-series. However, this might not yield the same results, as r and %REC do not scale linearly with each other).

So we can use radius as an outcome variable. What does radius mean? 

> The radius parameter r effectively specifies the interval within which two values are counted as being recurrent. 

So in other words, if we kept %REC = 5.0, that forces the radius to vary so that 5.0% of all possible points show recurrence. If the radius is high, that means the two time-series were "farther apart" and the higher radius was necessary in order to bring those two time-series into 5.0% recurrence. Conversely, a low radius indicates the two time-series do share more recurrence and thus a tighter radius still achieved 5.0% recurrence. 

Let's model that!

* Predictors: maingroup & direction
* Grouping variables (random intercepts): name & story
* Random slopes: hand (for name...so for each participant we allow the slope to vary for right vs. left)

It's actually a good thing we have CRQA results for both left and right hands. We know there is not a world of a difference between them, so we can "double" the dataset by adding "hand" as a random effects variable! 

```{r Radius plots and models, message=FALSE, warning=FALSE}

# Plot means
final_n <- final %>%
  select(name, maingroup, direction, story, hand) %>%
  distinct() %>%
  count(maingroup, direction, hand)

final_summary <- final %>%
  group_by(maingroup, direction, hand) %>%
  summarise(radius_mean = mean(radius),
            radius_sd = sd(radius)) %>%
  ungroup() %>%
  left_join(final_n, by = c("maingroup", "direction", "hand")) %>%
  mutate(radius_se = radius_sd/sqrt(n))
  
p_means <- final_summary %>%
  filter(hand == "right") %>%
  ggplot(aes(x = maingroup, y = radius_mean, color = direction, fill = direction)) +
  geom_point(position = position_dodge(width = 0.25), size = 2) +
  geom_errorbar(aes(ymin = radius_mean - radius_se, ymax = radius_mean + radius_se),
                position = position_dodge(width = 0.25),
                width = 0.2,
                size = 1) +
  scale_y_continuous(limits = c(22, 34)) +
  labs(x = "",
       y = "radius",
       title = "Radius For 5% Recurrence (Right Hand)") +
  guides(color = FALSE,
         fill = FALSE) +
  theme_ipsum_rc() +
  theme(panel.grid.minor.y = element_blank(),
        plot.title = element_text(color = 'white'))

p_means
p_means %>%
  ggsave(filename = 'tislr_figures/radius_means_plot.png',
         width = 9,
         height = 9,
         units = 'in',
         device = 'png')


# Fit model
m1 <- lmer(data = final, radius ~ maingroup * direction + (1 + hand|name) + (1|story))
tidy(m1, conf.int = TRUE) %>%
  mutate_if(is.numeric, round, 2) %>%
  filter(effect == 'fixed') %>%
  select(-effect, -group)

# Plot model coefficients
# https://cran.r-project.org/web/packages/sjPlot/vignettes/plot_model_estimates.html
p_coef <- plot_model(m1, 
           show.values = TRUE, value.offset = .3, 
           order.terms = c(4, 1:3, 5:7),
           width = .2,
           dot.size = 1,
           title = "Model Predictors for Radius (95% CI)",
           vline.color = "gray80") +
  scale_y_continuous(limits = c(-8, 7)) +
  theme_ipsum_rc() +
  theme(panel.grid.minor.x = element_blank(),
        plot.title = element_text(color = 'white'))

p_coef

p_coef %>%
  ggsave(filename = 'tislr_figures/radius_coefficient_plot.png',
         width = 12,
         height = 7,
         units = 'in',
         device = 'png')

tab_model(m1)
```

The interpretation is :  

1. We have a main effect of direction. RV results in smaller radius values, which means *more tightly coupled* viewing. Reversing video makes people pay more attention to the hands. 
1. There is no main effect of MainGroup...although Hearing Novice gets close. 
1. We have an interaction of Maingroup & Direction, where Hearing Noivce, when confronted w reverse video, become *less coupled* to hand position. It's possible they dis-engage and engage in poorer viewing behavior in general!


# Eye Gaze & Hand Position Plot

We need a plot like this for the poster, to give people an idea of what we are talking about. This was submitted with the abstract as figure 2. Re-creating it here. 

```{r Plot Hand and Eye Position, warning=FALSE, message=FALSE}
# Averaged eye gaze position for Three Bears
y_averaged <- read_csv("bears_time_series_group_averages.csv")

# Right hand position data 
y_rhand <- final %>%
  filter(story == "bears" & hand == "right") %>%
  slice(1) %>%
  pull(hand_y) %>%
  unlist() %>%
  as_tibble() %>%
  add_rownames() %>%
  rename(gaze_point_index = rowname,
         y = value) %>%
  mutate_at(vars(gaze_point_index), as.numeric)

# Left hand position data (Changed my mind, not plotting this.)
# y_lhand <- final %>%
#   filter(story == "bears" & hand == "left") %>%
#   slice(1) %>%
#   pull(hand_y) %>%
#   unlist() %>%
#   as_tibble() %>%
#   add_rownames() %>%
#   rename(gaze_point_index = rowname,
#          y = value) %>%
#   mutate_at(vars(gaze_point_index), as.numeric)


# Let's plot...
# First, get the eye gaze plot. 
p1 <- y_averaged %>%
  filter(maingroup == "DeafEarly" | maingroup == "HearingNovice") %>%
  ggplot() + 
  geom_line(aes(x = gaze_point_index, y = y_mean, color = maingroup), size = 1) +
  geom_ribbon(aes(x = gaze_point_index, 
                  ymin = y_mean - y_se, 
                  ymax = y_mean + y_se,
                  fill = maingroup), alpha = 0.2)

# Add right hand
p1 <- p1 +
  geom_step(data = y_rhand, aes(x = gaze_point_index, y = y),
            color = "#7CAE00", alpha = 0.7, size = 1) 
#  annotate("text", label = "Right Hand Position", x = 1950, y = 850, color = "#7CAE00")

# Refine!
p1 <- p1 +
  scale_y_reverse(limits = c(900,300)) +
  labs(x = "", 
       y = "", 
       title = "", 
       color = "", 
       fill = "") +
  guides(fill = FALSE,
         color = FALSE) +
 theme_linedraw() +
  theme(panel.grid.minor.y = element_blank(),
        legend.position = "top",
        panel.border = element_blank(),
        axis.ticks = element_blank())

p1 

p1 %>%
  ggsave(filename = 'tislr_figures/eye_gaze_hand_position.png',
         width = 12,
         height = 8,
         units = 'in',
         device = 'png')
```

# Participant N

How many participants per group, anyway? 

```{r Count N}
final %>%
  group_by(maingroup) %>%
  summarise(n = n_distinct(name))
```

# Entropy

r_ENTR is a normalized measure of entropy - closer to 1 = chaotic system. Closer to 0 = more predictable system. The following analysis suggests that HearingNovice have a more disordered looking relationship with hand, while other groups have more orderly looking behavior. 

Do we want this on the poster too? Kinda the same thing. I think no. Because it just adds to the story without much substance. Ppl will be like 'Entropy, what?'. 

```{r Entropy, message=FALSE, warning=FALSE}

# Plot means
final_n <- final %>%
  select(name, maingroup, direction, story, hand) %>%
  distinct() %>%
  count(maingroup, direction, hand)

final_summary_entr <- final %>%
  group_by(maingroup, direction, hand) %>%
  summarise(entr_mean = mean(r_ENTR),
            entr_sd = sd(r_ENTR)) %>%
  ungroup() %>%
  left_join(final_n, by = c("maingroup", "direction", "hand")) %>%
  mutate(entr_se = entr_sd/sqrt(n))
  
p_means_entr <- final_summary_entr %>%
  filter(hand == "right") %>%
  ggplot(aes(x = maingroup, y = entr_mean, color = direction, fill = direction)) +
  geom_point(position = position_dodge(width = 0.25), size = 2) +
  geom_errorbar(aes(ymin = entr_mean - entr_se, ymax = entr_mean + entr_se),
                position = position_dodge(width = 0.25),
                width = 0.2,
                size = 1) +
  scale_y_continuous(labels = scales::percent) +
  labs(x = "",
       y = "normalized Shannon entropy",
       title = "Normalized Entropy Values") +
  theme_ipsum_rc() +
  theme(panel.grid.minor.y = element_blank())

p_means_entr
p_means_entr %>%
  ggsave(filename = 'tislr_figures/entropy_means_plot.png',
         width = 12,
         height = 9,
         units = 'in',
         device = 'png')


# Fit model
m2 <- lmer(data = final, r_ENTR ~ maingroup * direction + (1 + hand|name) + (1|story))
tidy(m2, conf.int = TRUE) %>%
  mutate_if(is.numeric, round, 3) %>%
  filter(effect == 'fixed') %>%
  select(-effect, -group)

# Plot model coefficients
# https://cran.r-project.org/web/packages/sjPlot/vignettes/plot_model_estimates.html
p_coef_entr <- plot_model(m2, 
           show.values = TRUE, value.offset = .3, 
           order.terms = c(4, 1:3, 5:7),
           width = .2,
           dot.size = 1,
           vline.color = 'gray50',
           title = "Model Predictors for Normalized Entropy (95% CI)") +
  theme_ipsum_rc() +
  scale_y_continuous(limits = c(-0.3, 0.3))

p_coef_entr

p_coef_entr %>%
  ggsave(filename = 'tislr_figures/entropy_coefficient_plot.png',
         width = 12,
         height = 9,
         units = 'in',
         device = 'png')
```

# Best and Worst CRQA Plots

But we can use Entropy to let us know what the best and worst CRQA plots are. 

```{r CRQA Plots}
# Reuse optimizeParam mlpars
mlpar <- list(lgM = 120,
              radiusspan = 100,
              radiussample = 10,
              normalize = 2,
              rescale = 2,
              mindiagline = 2,
              minvertline = 2,
              tw = 0,
              whiteline = FALSE,
              recpt = FALSE,
              fnnpercent = 10,
              typeami = "maxlag")

# Reuse run_crqa function
run_crqa <- function(a, b, c, d, e){
  crqa(a, 
       b, 
       radius = c, 
       embed = d, 
       delay = e, 
       rescale = 2, 
       normalize = 2, 
       mindiagline = 2, 
       minvertline = 2, 
       tw = 0, 
       whiteline = FALSE, 
       recpt = FALSE, 
       side = 'both')
}

# Best CRQA (lowest entropy score)
best <- final %>%
  filter(hand == 'right') %>%
  filter(direction == 'fw') %>%
  arrange(r_ENTR) %>%
  slice(1)

# Worst CRQA (highest entropy score)
worst <- final %>%
  filter(hand == 'right') %>%
  filter(direction == 'rv') %>%
  arrange(desc(r_ENTR)) %>%
  slice(2)


# Get the recurrence plot by re-running optimizeParam & CRQA 
# best_param <- optimizeParam(best$eye_y[[1]], best$hand_y[[1]], mlpar)
# best_crqa <- run_crqa(best$eye_y[[1]], best$hand_y[[1]], best_param$radius, best_param$emddim, best_param$delay)

best_crqa <- run_crqa(best$eye_y[[1]], best$hand_y[[1]], best$radius, best$embeddim, best$delay)


# worst_param <- optimizeParam(worst$eye_y[[1]], worst$hand_y[[1]], mlpar)
# worst_crqa <- run_crqa(worst$eye_y[[1]], worst$hand_y[[1]], worst_param$radius, worst_param$emddim, worst_param$delay)

worst_crqa <- run_crqa(worst$eye_y[[1]], worst$hand_y[[1]], worst$radius, worst$embeddim, worst$delay)

# Plot
# params = list(unit = 10, labelx = "Time", labely = "Time", cols = "blue", pcex = 0.25)
# 
# best_plot <- plotRP(best_crqa$RP, params)
# worst_plot <- plotRP(worst_crqa$RP, params)


# Use geom_tile
vir_colors <- viridis::viridis(2, alpha = 1, begin = 0.2, end = 0.8, option = "D")

theme_black = function(base_size = 12, base_family = "") {
  theme_grey(base_size = base_size, base_family = base_family) %+replace%
    theme(
      # Specify axis options
      axis.line = element_blank(),  
      axis.text.x = element_blank(),  
      axis.text.y = element_blank(),  
      axis.ticks = element_blank(),  
      axis.title.x = element_text(size = base_size, color = "black", margin = margin(0, 10, 0, 0)),  
      axis.title.y = element_text(size = base_size, color = "black", angle = 90, margin = margin(0, 2, 0, 0)),  
      axis.ticks.length = unit(0.3, "lines"),   
      # Specify legend options
      # legend.background = element_rect(color = NA, fill = "black"),  
      # legend.key = element_rect(color = "white",  fill = "black"),  
      # legend.key.size = unit(1.2, "lines"),  
      # legend.key.height = NULL,  
      # legend.key.width = NULL,      
      # legend.text = element_text(size = base_size*0.8, color = "white"),  
      # legend.title = element_text(size = base_size*0.8, face = "bold", hjust = 0, color = "white"),  
      # legend.position = "right",  
      # legend.text.align = NULL,  
      # legend.title.align = NULL,  
      # legend.direction = "vertical",  
      # legend.box = NULL, 
      # Specify panel options
      panel.background = element_rect(fill = vir_colors[1], color  =  NA),  
      panel.border = element_rect(fill = NA, color = "black", size = 2),  
      panel.grid.major = element_blank(),  
      panel.grid.minor = element_blank(),  
      panel.spacing = unit(0.5, "lines"),
      # Specify facetting options
      # strip.background = element_rect(fill = "grey30", color = "grey10"),  
      # strip.text.x = element_text(size = base_size*0.8, color = "white"),  
      # strip.text.y = element_text(size = base_size*0.8, color = "white",angle = -90),  
      # Specify plot options
      plot.background = element_rect(color = "white", fill = "white"),  
      plot.title = element_text(size = base_size*1.2, color = "white"),  
      plot.margin = unit(rep(1, 4), "lines")
    )
}

best_rp <- summary(best_crqa$RP) %>%
  as_tibble() %>%
  ggplot(aes(x = i, y = j)) +
  geom_tile(fill = vir_colors[2]) +
  coord_fixed(expand = FALSE, ratio = 1) +
  theme_black() +
  labs(x = 'Eye Gaze Position',
       y = 'Hand Position')

worst_rp <- summary(worst_crqa$RP) %>%
  as_tibble() %>%
  ggplot(aes(x = i, y = j)) +
  geom_tile(fill = vir_colors[2]) +
  coord_fixed(expand = FALSE, ratio = 1) +
  theme_black() +
  labs(x = 'Eye Gaze Position',
       y = 'Hand Position')


# save_plot("tislr_figures/best_plot.png", best_plot)
# save_plot("tislr_figures/worst_plot.png", worst_plot)

# Make plots
best_rp %>%
  ggsave(filename = 'tislr_figures/best_rp.png',
         width = 9,
         height = 9,
         units = 'in',
         device = 'png')

worst_rp %>%
  ggsave(filename = 'tislr_figures/worst_rp.png',
         width = 9,
         height = 9,
         units = 'in',
         device = 'png')

# Print plot and info
best_rp

best %>%
  select(maingroup, story, direction, hand)

worst_rp

worst %>%
  select(maingroup, story, direction, hand)
```
